+++ sdk/eventhub/azure-eventhub/azure/eventhub/_common.py
@@ -60,8 +60,8 @@ from .amqp import (
 
 if TYPE_CHECKING:
     import datetime
-    MessageContent = TypedDict("MessageContent", {"content": bytes, "content_type": str})
 
+MessageContent = TypedDict("MessageContent", {"content": bytes, "content_type": str})
 PrimitiveTypes = Optional[Union[
     int,
     float,